organized src file structute. continuing work on Profile page. starting to bring in materialize components
