+++ src/components/Profile/Bio.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 
 const Bio = (props) => (
     <div>
