+++ src/components/Profile/Projects.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 
 const Projects = (props) => (
     <div>
