+++ src/components/Profile/Tasks.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react'
+import React from 'react'
 
 const Tasks = (props) => (
     <div>
