+++ src/containers/Access.js
@@ -1,7 +1,7 @@
-import React, {Component} from 'react'
+import React, { Component } from 'react'
 // import {Link} from 'react-router-dom'
-import Navbar from '../components/Navbar'
-import Footer from '../components/Footer'
+import Navbar from '../components/Recurrent/Navbar'
+import Footer from '../components/Recurrent/Footer'
 import Login from '../components/Login'
 
 // This is the Login page. Named "Access" to avoid
