+++ src/containers/Registration.js
@@ -1,9 +1,9 @@
 import React, {Component} from 'react'
 // import {Link} from 'react-router-dom'
-import Navbar from '../components/Navbar'
-import Footer from '../components/Footer'
+import Navbar from '../components/Recurrent/Navbar'
+import Footer from '../components/Recurrent/Footer'
 import SignUp from '../components/SignUp'
-import Login from '../components/Login'
+import Login from '../components/Login/Login'
 
 class Registration extends Component {
   render () {
